# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/namespaces/{namespace}/notebooks": {
            "get": {
                "description": "Get Notebooks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notebooks"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/output.Notebooks"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Notebook",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notebooks"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Notebook info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/input.Notebook"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/namespaces/{namespace}/notebooks/{notebook}": {
            "get": {
                "description": "Get Notebook",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notebooks"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Notebook",
                        "name": "notebook",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/output.Notebook"
                        }
                    }
                }
            }
        },
        "/namespaces/{namespace}/poddefaults": {
            "get": {
                "description": "Get pod defaults",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Defaults"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/output.PodDefaults"
                        }
                    }
                }
            }
        },
...
        "/namespaces/{namespace}/{notebook}": {
            "delete": {
                "description": "Delete Notebook",
                "tags": [
                    "Notebooks"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Notebook",
                        "name": "notebook",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "description": "Patch Notebook",
                "tags": [
                    "Notebooks"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Notebook",
                        "name": "notebook",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Notebook info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/input.StartStop"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "input.Datavolume": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string"
                },
                "extraFields": {
                    "$ref": "#/definitions/input.ExtraFields"
                },
                "mode": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "templatedName": {
                    "type": "string"
                },
                "templatedPath": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "input.ExtraFields": {
            "type": "object"
        },
        "input.Gpus": {
            "type": "object",
            "properties": {
                "num": {
                    "type": "string"
                },
                "vendor": {
                    "type": "string"
                },
                "vendors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "input.Notebook": {
            "type": "object",
            "properties": {
                "affinityConfig": {
                    "type": "string"
                },
                "allowCustomImage": {
                    "type": "boolean"
                },
                "configurations": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "cpu": {
                    "type": "string"
                },
                "cpuLimit": {
                    "type": "string"
                },
                "customImage": {
                    "type": "string"
                },
                "customImageCheck": {
                    "type": "boolean"
                },
                "datavols": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/input.Datavolume"
                    }
                },
                "gpus": {
                    "$ref": "#/definitions/input.Gpus"
                },
                "image": {
                    "type": "string"
                },
                "imagePullPolicy": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                },
                "memoryLimit": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "noWorkspace": {
                    "type": "boolean"
                },
                "serverType": {
                    "type": "string"
                },
                "shm": {
                    "type": "boolean"
                },
                "tolerationGroup": {
                    "type": "string"
                },
                "workspace": {
                    "$ref": "#/definitions/input.Datavolume"
                }
            }
        },
        "input.StartStop": {
            "type": "object",
            "properties": {
                "stopped": {
                    "type": "boolean"
                }
            }
        },
...
        "output.Gpus": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "output.Notebook": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "cpu": {
                    "type": "string"
                },
                "gpus": {
                    "$ref": "#/definitions/output.Gpus"
                },
                "image": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "serverType": {
                    "type": "string"
                },
                "shortImage": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/output.Status"
                },
                "volumes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "output.Notebooks": {
            "type": "object",
            "properties": {
                "notebooks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/output.Notebook"
                    }
                }
            }
        },
...
        "output.PodDefault": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                }
            }
        },
        "output.PodDefaults": {
            "type": "object",
            "properties": {
                "poddefaults": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/output.PodDefault"
                    }
                }
            }
        },
        "output.Status": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
    }
}